+++ README.md
@@ -1,0 +1,20 @@
+# Проект анализа финансовых данных
+
+Этот проект предназначен для анализа финансовых данных, включая расходы по картам, курсы валют и цены акций.
+
+## Запуск проекта
+
+1. Убедитесь, что у вас установлен Python 3.8 или выше.
+2. Установите зависимости:
+
+   pip install -r requirements.txt
+3. Создайте файл .env в корне проекта и добавьте туда API-ключ:
+API_KEY=ваш_api_ключ
+
+### 2. **Добавить `.env_example`**
+
+Создайте файл `.env_example` в корне проекта. Это шаблон для создания `.env`:
+
+```env
+# Ключ API для получения курсов валют
+API_KEY=ваш_api_ключ
